Change rating category ID to int

+++ src/main/java/com/klausapp/scorecalculator/dao/RatingCategory.java
@@ -2,5 +2,5 @@ package com.klausapp.scorecalculator.dao;
 
 import java.math.BigDecimal;
 
-public record RatingCategory(Long id, String name, BigDecimal weight) {
+public record RatingCategory(Integer id, String name, BigDecimal weight) {
 }
